Added CBA settings, moved configs to files
