+++ addons/safestart/RscDisplayGetReady.hpp
@@ -1,0 +1,15 @@
+class RscDisplayGetReady: RscDisplayMainMap {
+    class controls {
+        #include "BriefingScreenControls.hpp"
+    };
+};
+class RscDisplayClientGetReady: RscDisplayGetReady {
+    class controls {
+        #include "BriefingScreenControls.hpp"
+    };
+};
+class RscDisplayServerGetReady: RscDisplayGetReady {
+    class controls {
+        #include "BriefingScreenControls.hpp"
+    };
+};
